♻️ refactor: ~ModalView -> ~Modal
